requirements.txt
fastapi
uvicorn
tweepy
openai
apscheduler
python-dotenv
aiofiles
